requirements.txt
bs4==0.0.1
requests==2.31
tqdm==4.64.1
pyyaml==6.0.1